Added scoreHandler to make score time based as of now, with decrements for every collision with enemy

+++ src/h/scoreComponent.h
@@ -6,13 +6,13 @@
 
 class ScoreComponent : public Component {
 private:
-    int score;
+    float score;
     float multiplier;
 
 public:
     ScoreComponent(Entity* entity);
     ~ScoreComponent();
-    void addScore(int scoreIncrease);
+    void addScore(float scoreIncrease);
     int getScore();
     void resetScore();
     void setMultiplier(float multiplier);
